Move sync button to global top bar header
Centralizes the Tacticus API sync button (with auto-sync-on-mount)
into a shared SyncButton component in the authenticated header,
removing ~30 lines of duplicated sync logic from each of the 4
authenticated pages.
